Fast detection of normally closed windows.
Windows when closed normally receive a window "unload" event.  A
window isn't able to write to the database at this point (it will
receive no more events in the event loop), but it is able to write to
local storage and/or send a message to the shared web worker.

This isn't 100% reliable (windows can crash without having a chance to
process the unload event), but it does provide for a fast path for
detecting dead windows when they've been closed normally.  Pings are
still used to detect crashed windows.